Refactored logic into base class.
Added reverse geocoder class
