+++ libs/common/utils/geometry/geoprocessing/src/lib/common-utils-geometry-geoprocessing.ts
@@ -1,3 +1,4 @@
 export * from './core/types';
 
-export { Geocoder } from './geocoding/geocoding';
+export { Geocoder } from './geocoder/geocoder';
+export { ReverseGeocoder } from './reverse-geocoder/reverse-geocoder';
